Incorporated modification from Julia Deneva's newest version of psrfits2psrfits.
These changes include bug fixes and new features:
--The bugs Slavko found are fixed: (1) the last (partial) row of an
observation now gets dropped appropriately, including when the last
file of an observation contains only one row, and (2) in some cases,
the last row of *every* file in the obs. used to get dropped, that is
fixed too.

--Changed the way new files are created. I now create a template from
the 1st input file and modify it to the correct table width and column
types for the desired output. This template is then used in creating
the output files for that observation.

--New command-line option letting the user specify a size for the
output files in GB. I've set the default to 10GB.

--Removed the band inversion command-line option because that's not
implemented yet (no demand for it so far).